updated the folder structure and overall refactored to match main branch.
