+++ src/pages/index.tsx
@@ -1,13 +1,13 @@
 import Box from "@mui/material/Box";
 import React from "react";
-import Video from "../wrappers/HomePage/Video";
+import Video from "../wrappers/Video";
 import { HeadFC } from "gatsby";
 import Layout from "../components/Layout";
-import Events from "../wrappers/HomePage/Events";
+import Events from "../wrappers/Events";
 import DrawerAppBar from "../components/navbar/Navbar.react";
-import Projects from "../wrappers/HomePage/Projects";
-import Volunteers from "../wrappers/HomePage/Volunteers";
-import Testimonial from "../wrappers/HomePage/Testimonial";
+import Projects from "../wrappers/Projects";
+import Volunteers from "../wrappers/Volunteers";
+import Testimonial from "../wrappers/Testimonial";
 
 const IndexPage = (props: any) => {
   return (
